Update Mapland composed key to use uint256 instead of uint32

+++ packages/contracts/mud.config.ts
@@ -4,8 +4,8 @@ export default mudConfig({
   tables: {
     MapLand: {
       keySchema: {
-        x: "uint32",
-        y: "uint32",
+        x: "uint256",
+        y: "uint256",
       },
       schema: {
         tokenAddress: "address",
